poc/retrieval_benchmark/pyproject.toml
[project]
name = "retrieval-benchmark"
version = "0.1.0"
description = "Benchmark framework for comparing document retrieval strategies"
requires-python = ">=3.11"
dependencies = [
    # Vector stores
    "chromadb>=0.4.0",
    "numpy>=1.26.0",
    
    # Embeddings
    "sentence-transformers>=2.2.0",
    
    # LLM (for Phase 2)
    "ollama>=0.4.0",
    
    # Config & CLI
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.0",
    
    # Utils
    "rich>=13.0",
    "pandas>=2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[project.scripts]
retrieval-benchmark = "retrieval_benchmark.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
